Separate db counter
