+++ lib/db/stats.js
@@ -1,6 +1,6 @@
-import {createIndex, findAll, upsertOne} from './mongo.js'
+import {createIndex, findAll} from './mongo.js'
 import {getDay} from './query.js'
-import {encodePage, decodePage} from '../core/format.js'
+import {decodePage} from '../core/format.js'
 
 
 await init()
@@ -14,26 +14,6 @@ export async function readLatestSites() {
 	const day = getDay()
 	const all = await findAll('sites', {day}, {site: 1, total: 1})
 	return all.map(stats => ({key: stats.site, value: stats.total}))
-}
-
-export async function storeCounter(counter) {
-	if (!counter.site || !counter.page) {
-		console.log('Missing site or page')
-		return
-	}
-	const update = {total: 1}
-	for (const field in counter.stats) {
-		const value = counter.stats[field]
-		if (value) {
-			const key = `${field}:${value}`
-			update[key] = 1
-		}
-	}
-	const page = encodePage(counter.page)
-	const pageKey = `page:${page}`
-	const updateWithPage = {...update, [pageKey]: 1}
-	await upsertOne('sites', {site: counter.site, day: counter.day}, {$inc: updateWithPage})
-	await upsertOne('pages', {site: counter.site, day: counter.day, page: counter.page}, {$inc: update})
 }
 
 export async function readSiteStats(query) {
