+++ lib/server/counter.js
@@ -1,5 +1,5 @@
 import {promises as fs} from 'fs'
-import {storeCounter} from '../db/stats.js'
+import {storeCounter} from '../db/counter.js'
 import {Counter} from '../core/counter.js'
 
 const logo = await fs.readFile('public/librecounter.svg')
